Move *Requests views to extrequests application
WorkshopRequest, EventRequest, EventSubmission,
DCSelfOrganizedEventRequest, ProfileUpdateRequest, and
InvoiceRequest views were moved from `workshops` to `extrequests`.

New ChangeRequestState class-based-view was added to keep things simple.
